+++ src/components/OrderMenu/orderMenu.css
@@ -1,0 +1,1 @@
+#order-menu-container .btn-top{margin-top:20px}#order-menu-container h5{margin:auto auto 1em}#order-menu-container input{text-align:right}#order-menu-container .buy{--color: #20df20}#order-menu-container .sell{--color: #df2020}
