--- src/components/PendingOrderMenu/pendingOrder.css
@@ -1,1 +1,0 @@
-#container-pending-order h5{margin:auto auto 1em}#container-pending-order input{text-align:right}#container-pending-order .buy{--color: #20df20}#container-pending-order .sell{--color: #df2020}
